+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+- Removed old parent directory creation logic for checkpoint file
+
 ## [0.3.0] - 2024-11-23
 - Explicitly display "CLIP" in progress bar when gradient norm is clipped
 - Do not create parent directory for checkpoint file until save is called