+++ CHANGELOG.md
@@ -4,10 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.3.1] - 2024-11-27
 - Removed old parent directory creation logic for checkpoint file
 - Added dropout option to activated block
 - Toggle loss between train and eval in PyTorch training loop
+- Added cross-entropy loss with label smoothing on and off for train and eval
 
 ## [0.3.0] - 2024-11-23
 - Explicitly display "CLIP" in progress bar when gradient norm is clipped